+++ src/main/java/pl/asie/computronics/reference/Names.java
@@ -20,6 +20,7 @@ public class Names {
 		FSP_SteamTransporter = "steam_transporter",
 		FZ_ChargeConductor = "charge_conductor",
 		MFR_DSU = "deep_storage_unit",
+		OpenBlocks_BuildingGuide = "building_guide",
 		Railcraft_BoilerFirebox = "boiler_firebox",
 		Railcraft_DigitalDetector = "digital_detector",
 		Railcraft_DigitalReceiverBox = "digital_receiver_box",